Nagar-AI/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0